+++ basis_functions.py
@@ -1,29 +1,30 @@
+#!/usr/bin/env python
 """
-This script is used to generate basis sources for the 
-kCSD method Jan et.al (2012) for 3D case.
-
-These scripts are based on Grzegorz Parka's, 
-Google Summer of Code 2014, INFC/pykCSD  
-
+This script is used to generate basis sources for the
+kCSD method Jan et.al (2012) for 1D,2D and 3D cases.
+Two 'types' are described here, gaussian and step source,
+These can be easily extended.
+These scripts are based on Grzegorz Parka's,
+Google Summer of Code 2014, INFC/pykCSD
 This was written by :
-Michal Czerwinski, Chaitanya Chintaluri  
+Michal Czerwinski, Chaitanya Chintaluri
 Laboratory of Neuroinformatics,
 Nencki Institute of Experimental Biology, Warsaw.
 """
+from __future__ import division
+
 import numpy as np
 
 def gauss(d, stdev, dim):
     """Gaussian function
-
     Parameters
     ----------
     d : floats or np.arrays
-        Distance array to the point of evaluation 
+        Distance array to the point of evaluation
     stdev : float
         cutoff range
     dim : int
         dimension of the gaussian function
-
     Returns
     -------
     Z : floats or np.arrays
@@ -34,32 +35,28 @@ def gauss(d, stdev, dim):
 
 def step_1D(d, R):
     """Returns normalized 1D step function.
-
     Parameters
     ----------
     d : floats or np.arrays
-        Distance array to the point of evaluation 
+        Distance array to the point of evaluation
     R : float
         cutoff range
-
     Returns
     -------
     s : Value of the function (d  <= R) / R
     """
     s = (d  <= R)
     s = s / R #normalize with width
-    return s        
+    return s
 
 def gauss_1D(d, three_stdev):
     """Returns normalized gaussian 2D scale function
-
     Parameters
     ----------
     d : floats or np.arrays
         Distance array to the point of evaluation
     three_stdev : float
         3 * standard deviation of the distribution
-
     Returns
     -------
     Z : (three_std/3)*(1/2*pi)*(exp(-0.5)*stddev**(-2) *(d**2))
@@ -70,17 +67,15 @@ def gauss_1D(d, three_stdev):
 
 def gauss_lim_1D(d, three_stdev):
     """Returns gausian 2D function cut off after 3 standard deviations.
-
     Parameters
     ----------
     d : floats or np.arrays
         Distance array to the point of evaluation
     three_stdev : float
         3 * standard deviation of the distribution
-
     Returns
     -------
-    Z : (three_std/3)*(1/2*pi)*(exp(-0.5)*stddev**(-2) *((x-mu)**2)), 
+    Z : (three_std/3)*(1/2*pi)*(exp(-0.5)*stddev**(-2) *((x-mu)**2)),
         cut off = three_stdev
     """
     Z = gauss_1D(d, three_stdev)
@@ -89,31 +84,28 @@ def gauss_lim_1D(d, three_stdev):
 
 def step_2D(d, R):
     """Returns normalized 2D step function.
-
     Parameters
     ----------
     d : float or np.arrays
         Distance array to the point of evaluation
     R : float
         cutoff range
-    
+
     Returns
     -------
     s : step function
     """
     s = (d <= R) / (np.pi*(R**2))
-    return s        
+    return s
 
 def gauss_2D(d, three_stdev):
     """Returns normalized gaussian 2D scale function
-
     Parameters
     ----------
     d : floats or np.arrays
          distance at which we need the function evaluated
     three_stdev : float
         3 * standard deviation of the distribution
-
     Returns
     -------
     Z : function
@@ -125,14 +117,12 @@ def gauss_2D(d, three_stdev):
 
 def gauss_lim_2D(d, three_stdev):
     """Returns gausian 2D function cut off after 3 standard deviations.
-
     Parameters
     ----------
     d : floats or np.arrays
          distance at which we need the function evaluated
     three_stdev : float
         3 * standard deviation of the distribution
-
     Returns
     -------
     Z : function
@@ -143,14 +133,12 @@ def gauss_lim_2D(d, three_stdev):
 
 def gauss_3D(d, three_stdev):
     """Returns normalized gaussian 3D scale function
-
     Parameters
     ----------
     d : floats or np.arrays
-        distance at which we need the function evaluated        
+        distance at which we need the function evaluated
     three_stdev : float
         3 * standard deviation of the distribution
-
     Returns
     -------
     Z : funtion
@@ -162,14 +150,12 @@ def gauss_3D(d, three_stdev):
 
 def gauss_lim_3D(d, three_stdev):
     """Returns normalized gaussian 3D scale function cut off after 3stdev
-
     Parameters
     ----------
     d : floats or np.arrays
-        distance at which we need the function evaluated        
+        distance at which we need the function evaluated
     three_stdev : float
         3 * standard deviation of the distribution
-
     Returns
     -------
     Z : funtion
@@ -181,19 +167,17 @@ def gauss_lim_3D(d, three_stdev):
 
 def step_3D(d, R):
     """Returns normalized 3D step function.
-
     Parameters
     ----------
     d : floats or np.arrays
-        distance at which we need the function evaluated        
+        distance at which we need the function evaluated
     R : float
         cutoff range
-
     Returns
     -------
     s : step function in 3D
     """
-    
+
     s = 3/(4*np.pi*R**3)*(d <= R)
     return s
